requirements.txt
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.11.2

numpy
pandas
xlrd
tensorflow-cpu
transformers

